style : 이름변경
시스템 상수로 이름을 변경하였습니다.
